move switch from edit profile modal to profile header
