+++ client/src/components_volunteer/EditAccountInfoModal.js
@@ -66,32 +66,6 @@ export default function EditAccountInfoModal(props) {
   ];
   const timeNames = ["Morning", "Afternoon", "Evening", "Weekdays", "Weekends"];
   const languages = ["English", "Spanish", "Mandarin", "Cantonese", "Other"];
-  const [allowSMS, setAllowSMS] = useState(false);
-
-  const handleChangeSMSPreference = (event) => {
-    event.persist();
-    setAllowSMS(!allowSMS);
-  };
-  const SMSNotificationSwitch = (
-    <Form.Group
-      controlId="preverify"
-      bssize="large"
-      style={{ marginLeft: 12, marginBottom: 8 }}
-    >
-      <Form.Check
-        type="switch"
-        id="custom-switch"
-        style={{ color: "#7F7F7F", fontSize: 14 }}
-        label={
-          allowSMS
-            ? "Covaid will text you when you receive new requests"
-            : "You will NOT be receiving texts from Covaid"
-        }
-        checked={allowSMS}
-        onChange={handleChangeSMSPreference}
-      />
-    </Form.Group>
-  );
 
   const setCurrentUserObject = (userList, fullList, setFunction) => {
     for (var i = 0; i < fullList.length; i++) {
@@ -295,7 +269,6 @@ export default function EditAccountInfoModal(props) {
         coordinates: latlong,
       },
       languages: selectedLanguages,
-      allowSMS: allowSMS,
     };
 
     if (showChangeAssocModal) {
@@ -357,7 +330,7 @@ export default function EditAccountInfoModal(props) {
           setPhone(props.user.phone);
           setLatLong(props.user.latlong);
           getZip(props.user.latlong);
-          setAllowSMS(props.user.allowSMS);
+          // setAllowSMS(props.user.allowSMS);
           setAssociation(props.user.association);
           setAssociationName(props.user.association_name);
           setHasCar(props.user.offer.car);
@@ -510,7 +483,6 @@ export default function EditAccountInfoModal(props) {
                     <Form.Label id="regular-text-bold">
                       {props.translations[props.language].phone}
                     </Form.Label>
-                    {SMSNotificationSwitch}
                   </div>
 
                   <PhoneNumber
